requirements.txt
# 필수 데이터/모델/머신러닝/딥러닝 패키지
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
lightgbm>=3.3.0
joblib>=1.1.0

# 딥러닝 및 트랜스포머
torch>=1.13.0
transformers>=4.30.0
peft>=0.4.0
bitsandbytes>=0.41.0

# 텍스트 처리/유틸
tqdm>=4.62.0
pyyaml>=6.0

# KenLM 파이썬 바인딩 (pip 설치)
kenlm

# 하이퍼파라미터 튜닝
optuna>=3.0.0

# 기타(윈도우 호환, 경고 방지 등)
accelerate>=0.21.0

# (선택) wandb, matplotlib 등 실험 관리/시각화
# wandb>=0.15.0
# matplotlib>=3.5.0

python-dotenv>=1.0.0